requirements.txt
altair==5.5.0
annotated-types==0.7.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
duckdb==1.0.0
gitdb==4.0.12
GitPython==3.1.45
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multimethod==1.10
mypy_extensions==1.1.0
narwhals==2.5.0
numpy==2.0.2
packaging==24.2
pandas==2.2.2
pandera==0.20.3
pillow==10.4.0
plotly==5.22.0
protobuf==5.29.5
pyarrow==16.1.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
rich==13.7.1
rpds-py==0.27.1
six==1.17.0
smmap==5.0.2
streamlit==1.36.0
tenacity==8.5.0
toml==0.10.2
tornado==6.5.2
typeguard==4.4.4
typing-inspect==0.9.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wrapt==1.17.3
zipp==3.23.0
